property card grid view refactoring for constant values.
