+++ lib/constants/constants.dart
@@ -1,10 +1,46 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:intl/intl.dart';
 
 const kPrimaryColor = Color.fromRGBO(243, 68, 81, 1);
+const kSecondaryColor = Color.fromRGBO(243, 93, 67, 1);
+const kIconBackgroundColor = Color.fromRGBO(254, 237, 238, 1);
+const kTextColor = Color(0xFF586167);
 const kAppBarPrimaryColor = Color.fromRGBO(255, 255, 255, 0);
 const kAppBarMenuItems = ["home", "property", "project", "services", "city"];
-const kSecondaryColor = Color.fromRGBO(243, 93, 67, 1);
+const kDefaultSizedBoxHeight = 10.0;
+const kDefaultSizedBoxWidth = 10.0;
+const kElevation = 10.0;
+const kdropDownTitleStyle = 14.0;
+
+DateFormat kDateformat = DateFormat('MMMM dd, yyyy');
+
+// ************ Text *****************
+
+TextStyle kTextDefaultStyle = GoogleFonts.montserrat(
+    fontSize: kdropDownTitleStyle,
+    fontWeight: FontWeight.w500,
+    color: kTextColor);
+
+TextStyle kTextKeyStyle =
+    kTextDefaultStyle.copyWith(fontWeight: FontWeight.w600);
+
+TextStyle kTextHeader1Style =
+    GoogleFonts.montserrat(fontWeight: FontWeight.w700, fontSize: 24);
+
+TextStyle kTextHeader2Style =
+GoogleFonts.montserrat(fontWeight: FontWeight.w700, fontSize: 19);
+
+TextStyle kButtonStyle =
+    GoogleFonts.montserrat(color: Colors.white, fontSize: 17);
+
+// ************ Padding *****************
+
+const EdgeInsetsGeometry kDefaultPadding = EdgeInsets.all(8.0);
+
+// ************ SizedBox *****************
+const kDefaultSizedBox = SizedBox(height: kDefaultSizedBoxHeight);
+
 
 const menus = {
   "home": [],
@@ -14,7 +50,7 @@ const menus = {
   "city": ["Nashik", "Igatpuri"]
 };
 
-const menuList = ["home", "property", "about us", "city" ];
+const menuList = ["home", "property", "about us", "city"];
 
 const propertySearch = {
   "Property Type": ["Residential", "Commercial", "Plot", "Industrial"]
@@ -27,14 +63,3 @@ const maxRoomsDD = {
   "Plot": ["Agriculture", "Industrial", "NA", "Resort NA", "Amenity"],
   "Industrial": ["Factory", "Shed"]
 };
-
-TextStyle kMenuItemStyle = GoogleFonts.rubik(fontSize: 15, fontWeight: FontWeight.w500, color: kTextColor);
-const kIconColor = Color.fromRGBO(243, 68, 81, 1);
-const kIconBackgroundColor = Color.fromRGBO(254, 237, 238, 1);
-const kTextColor = Color(0xFF586167);
-TextStyle kTitleStyle = GoogleFonts.montserrat(fontWeight: FontWeight.w700, fontSize: 24);
-const kdropDownTitleStyle = 14.0;
-TextStyle khomePagePropertySearchPanelDDPanel =
-    GoogleFonts.rubik(fontSize: kdropDownTitleStyle, color: Colors.black);
-
-TextStyle buttonStyle = GoogleFonts.rubik(color: Colors.white, fontSize: 17);
