+++ lib/constants/controllers.dart
@@ -4,7 +4,7 @@ import 'package:the_asset_zone_web/controllers/search_controller.dart';
 import 'package:the_asset_zone_web/controllers/single_page_property_controller.dart';
 
 NavBarController navBarController = NavBarController.instance;
-PropertyController propertyDetailsFirestore = PropertyController.instance;
+PropertyController propertyController = PropertyController.instance;
 SearchController searchController = SearchController.instance;
 SinglePagePropertyController singlePagePropertyController =
     SinglePagePropertyController();
