+++ lib/screens/home/components/property_search_tablet_view.dart
@@ -44,7 +44,7 @@ class PropertySearchTabletView extends StatelessWidget {
                               padding: const EdgeInsets.all(15.0),
                               child: Text(
                                 "Budget ",
-                                style: khomePagePropertySearchPanelDDPanel,
+                                style: kTextDefaultStyle,
                               ),
                             ),
                             PropertySearchCardSearchRangeSliderSelectedValue(
